Added buttons to the UI that allow the user to go to a play mode, analyze mode, or a mode where you can change the position. Moved some of the functions to Functions.cpp to improve readability.

+++ src/PGN.cpp
@@ -133,7 +133,7 @@ namespace ChessGame {
 	std::string PGN::generatePGN(MoveNode* root, ChessBoard chessBoard, int index) {
 		std::string pgn_ans = "";
 		std::string move_pgn;
-		int moveNumber = root->moveNumber;
+		size_t moveNumber = root->moveNumber;
 		bool priorityLine = false;
 
 		if (root->move) {
